refactor!(node-stdlib): Rename command module to cmd
