+++ packages/node-stdlib/src/cmd/lookpath.ts
