+++ packages/node-stdlib/src/cmd/mod.ts
@@ -1,4 +1,4 @@
 // Copyright (c) 2020 Christopher Szatmary <cs@christopherszatmary.com>
 // All rights reserved. MIT License.
 
-export * from "./command";
+export * from "./lookpath";
