+++ packages/node-stdlib/src/index.ts
@@ -1,7 +1,7 @@
 // Copyright (c) 2020 Christopher Szatmary <cs@christopherszatmary.com>
 // All rights reserved. MIT License.
 
-export * as command from "./command/mod";
+export * as cmd from "./cmd/mod";
 export * as core from "./core/mod";
 export * as errors from "./errors/mod";
 export * as fatal from "./fatal/mod";
